NH3 ammonia network + working MACE ML potential on GB10 GPU
- network.py: ammonia (NO reduction to NH3) network, now the default. Rooted at
  adsorbed NO; dissociation NO->N+O, hydrogenation chain N+H->NH->NH2->NH3, and
  the associative NO+H->HNO/NOH fork. build_network dispatch: ammonia|branching|
  oxidation. Config defaults target=NH3, network=ammonia.
- calculators.py: mace backend uses default_dtype=float64 (geometry-opt accuracy).
- GB10/Blackwell (sm_121) GPU fix: torch cu128 wheel's nvrtc 12.8 rejects sm_121;
  upgrading nvidia-cuda-nvrtc-cu12>=12.9 (soname libnvrtc.so.12 stable) makes
  torch/MACE JIT work on GPU. Pinned in the `mace` extra; documented in USAGE.
- examples/no_to_nh3_pd.yaml; test for ammonia network reaching NH3 (32 tests).

Verified: MACE-MP-0 runs through the pipeline on the GB10 GPU (NO/Pd
E=-70.119 eV, matches CPU), real ML-potential energies/forces (not EMT).

+++ src/atosim/calculators.py
@@ -37,7 +37,9 @@ def make_calculator(cfg: MLIPConfig) -> "Calculator":
     if backend == "mace":
         from mace.calculators import mace_mp  # type: ignore
 
-        return mace_mp(model=cfg.model or "medium", device=cfg.device)
+        # float64 is recommended for geometry optimization / NEB (vs float32 MD).
+        return mace_mp(model=cfg.model or "medium", device=cfg.device,
+                       default_dtype="float64")
     if backend == "fairchem":
         # fairchem >=1.0 exposes an ASE calculator factory; kept lazy on purpose.
         from fairchem.core import OCPCalculator  # type: ignore
